--- lib/base_dataset.py
@@ -1,75 +1,0 @@
-#!/usr/bin/python
-# -*- encoding: utf-8 -*-
-
-import os
-import os.path as osp
-import json
-
-import torch
-from torch.utils.data import Dataset, DataLoader
-import torch.distributed as dist
-import cv2
-import numpy as np
-
-from lib.sampler import RepeatedDistSampler
-
-
-
-class BaseDataset(Dataset):
-    '''
-    '''
-    def __init__(self, dataroot, annpath, trans_func=None, mode='train'):
-        super(BaseDataset, self).__init__()
-        assert mode in ('train', 'val', 'test')
-        self.mode = mode
-        self.trans_func = trans_func
-
-        self.lb_ignore = -100
-        self.lb_map = None
-
-        with open(annpath, 'r') as fr:
-            pairs = fr.read().splitlines()
-        self.img_paths, self.lb_paths = [], []
-        for pair in pairs:
-            imgpth, lbpth = pair.split(',')
-            self.img_paths.append(osp.join(dataroot, imgpth))
-            self.lb_paths.append(osp.join(dataroot, lbpth))
-
-        assert len(self.img_paths) == len(self.lb_paths)
-        self.len = len(self.img_paths)
-
-    def __getitem__(self, idx):
-        impth, lbpth = self.img_paths[idx], self.lb_paths[idx]
-        img, label = self.get_image(impth, lbpth)
-        if not self.lb_map is None:
-            label = self.lb_map[label]
-        im_lb = dict(im=img, lb=label)
-        if not self.trans_func is None:
-            im_lb = self.trans_func(im_lb)
-        im_lb = self.to_tensor(im_lb)
-        img, label = im_lb['im'], im_lb['lb']
-        return img.detach(), label.unsqueeze(0).detach()
-
-    def get_image(self, impth, lbpth):
-        img = cv2.imread(impth)[:, :, ::-1].copy()
-        label = cv2.imread(lbpth, 0)
-        return img, label
-
-    def __len__(self):
-        return self.len
-
-
-if __name__ == "__main__":
-    from tqdm import tqdm
-    from torch.utils.data import DataLoader
-    ds = CityScapes('./data/', mode='val')
-    dl = DataLoader(ds,
-                    batch_size = 4,
-                    shuffle = True,
-                    num_workers = 4,
-                    drop_last = True)
-    for imgs, label in dl:
-        print(len(imgs))
-        for el in imgs:
-            print(el.size())
-        break
