+++ tools/train.py
@@ -19,7 +19,7 @@ from torch.utils.data import DataLoader
 
 from lib.models import model_factory
 from configs import set_cfg_from_file
-from lib.get_dataloader import get_data_loader
+from lib.data import get_data_loader
 from tools.evaluate import eval_model
 from lib.ohem_ce_loss import OhemCELoss
 from lib.lr_scheduler import WarmupPolyLrScheduler
